cleans up FBA print output.
..as desktop land may be confused by the "margin" of column 1.  Also removes page numbers.
Included for those using TK4- (and similar) on Hercules that don't have OPTION COPY or FIELDS=COPY alternatives.
Yes it is also trivial to do this with grep and cut (maybe awk). 
The exits are largely for RC 0000 as otherwise they are "assumed" here and there. This predates SECTIONs that would scope them.
STOP RUN could go in 30-FINISH, and this could be called upon an initial AT END condition.
See TSO BLK3390 and BLK3380 etc. for optimal block sizes.